// src/types/NodeConfig.generated.ts
export type NodeConfig = {
// ...
  'node'?: {
    'avail'?: {
// ...
      'fallback-s3-service-config'?: {
        /** S3 access key */
        'access-key'?: string;
        /** S3 bucket */
        'bucket'?: string;
        /** discard data after its expiry timeout */
        'discard-after-timeout'?: boolean;
        /** enable storage/retrieval of sequencer batch data from an AWS S3 bucket */
        'enable'?: boolean;
        /** prefix to add to S3 objects */
        'object-prefix'?: string;
        /** S3 region */
        'region'?: string;
        /** S3 secret key */
        'secret-key'?: string;
      };
      /** Avail chain wallet seed */
      'seed'?: string;
      /** timeout for batch inclusion on block finalisation over Avail chain (default 1m40s) */
      'timeout'?: string;
    };
// ...
    'data-availability'?: {
      /** enable Anytrust Data Availability mode */
      'enable'?: boolean;
      /** whether the Data Availability Service should fail immediately on errors (not recommended) */
      'panic-on-error'?: boolean;
      /** parent chain RPC connection attempts (spaced out at least 1 second per attempt, 0 to retry infinitely), only used in standalone daserver; when running as part of a node that node's parent chain configuration is used (default 15) */
      'parent-chain-connection-attempts'?: number;
      /** URL for parent chain node, only used in standalone daserver; when running as part of a node that node's L1 configuration is used */
      'parent-chain-node-url'?: string;
      /** Data Availability Service timeout duration for Store requests (default 5s) */
      'request-timeout'?: string;
      'rest-aggregator'?: {
        /** enable retrieval of sequencer batch data from a list of remote REST endpoints; if other DAS storage types are enabled, this mode is used as a fallback */
        'enable'?: boolean;
        /** number of stats entries (latency and success rate) to keep for each REST endpoint; controls whether strategy is faster or slower to respond to changing conditions (default 20) */
        'max-per-endpoint-stats'?: number;
        /** a URL to a list of URLs of REST das endpoints that is checked at startup; additive with the url option */
        'online-url-list'?: string;
        /** time interval to periodically fetch url list from online-url-list (default 1h0m0s) */
        'online-url-list-fetch-interval'?: string;
        'simple-explore-exploit-strategy'?: {
          /** number of consecutive GetByHash calls to the aggregator where each call will cause it to select from REST endpoints in order of best latency and success rate, before switching to explore mode (default 1000) */
          'exploit-iterations'?: number;
          /** number of consecutive GetByHash calls to the aggregator where each call will cause it to randomly select from REST endpoints until one returns successfully, before switching to exploit mode (default 20) */
          'explore-iterations'?: number;
        };
        /** strategy to use to determine order and parallelism of calling REST endpoint URLs; valid options are 'simple-explore-exploit' (default "simple-explore-exploit") */
        'strategy'?: string;
        /** how frequently to update the strategy with endpoint latency and error rate data (default 10s) */
        'strategy-update-interval'?: string;
        'sync-to-storage'?: {
          /** time to wait if encountered an error before retrying (default 1s) */
          'delay-on-error'?: string;
          /** eagerly sync batch data to this DAS's storage from the rest endpoints, using L1 as the index of batch data hashes; otherwise only sync lazily */
          'eager'?: boolean;
          /** when eagerly syncing, start indexing forward from this L1 block. Only used if there is no sync state */
          'eager-lower-bound-block'?: number;
          /** log only on failures to write when syncing; otherwise treat it as an error (default true) */
          'ignore-write-errors'?: boolean;
          /** when eagerly syncing, max l1 blocks to read per poll (default 100) */
          'parent-chain-blocks-per-read'?: number;
          /** period to request storage to retain synced data (default 360h0m0s) */
          'retention-period'?: string;
          /** directory to store the sync state in, ie the block number currently synced up to, so that we don't sync from scratch each time */
          'state-dir'?: string;
          /** sync even data that is expired; needed for mirror configuration (default true) */
          'sync-expired-data'?: boolean;
        };
        /** list of URLs including 'http://' or 'https://' prefixes and port numbers to REST DAS endpoints; additive with the online-url-list option */
        'urls'?: string[];
        /** time to wait until trying the next set of REST endpoints while waiting for a response; the next set of REST endpoints is determined by the strategy selected (default 2s) */
        'wait-before-try-next'?: string;
      };
      'rpc-aggregator'?: {
        /** Number of assumed honest backends (H). If there are N backends, K=N+1-H valid responses are required to consider an Store request to be successful. */
        'assumed-honest'?: number;
        /** enable storage of sequencer batch data from a list of RPC endpoints; this should only be used by the batch poster and not in combination with other DAS storage types */
        'enable'?: boolean;
        /** maximum HTTP POST body size to use for individual batch chunks, including JSON RPC overhead and an estimated overhead of 512B of headers (default 524288) */
        'max-store-chunk-body-size'?: number;
      };
      /** parent chain address of SequencerInbox contract */
      'sequencer-inbox-address'?: string;
    };
// ...
  };
  'parent-chain'?: {
    'blob-client'?: {
      /** Value to send with the HTTP Authorization: header for Beacon REST requests, must include both scheme and scheme parameters */
      'authorization'?: string;
      /** Beacon Chain RPC URL to use for fetching blobs (normally on port 3500) */
      'beacon-url'?: string;
      /** Full path of the directory to save fetched blobs */
      'blob-directory'?: string;
      /** Backup beacon Chain RPC URL to use for fetching blobs (normally on port 3500) when unable to fetch from primary */
      'secondary-beacon-url'?: string;
    };
    'connection'?: {
      /** limit size of arguments in log entries (default 2048) */
      'arg-log-limit'?: number;
      /** how long to wait for initial connection (default 1m0s) */
      'connection-wait'?: string;
      /** path to file with jwtsecret for validation - ignored if url is self or self-auth */
      'jwtsecret'?: string;
      /** number of retries in case of failure(0 mean one attempt) (default 2) */
      'retries'?: number;
      /** delay between retries */
      'retry-delay'?: string;
      /** Errors matching this regular expression are automatically retried */
      'retry-errors'?: string;
      /** per-response timeout (0-disabled) (default 1m0s) */
      'timeout'?: string;
      /** url of server, use self for loopback websocket, self-auth for loopback with authentication */
      'url'?: string;
      /** websocket message size limit used by the RPC client. 0 means no limit (default 268435456) */
      'websocket-message-size-limit'?: number;
    };
    /** if set other than 0, will be used to validate database and L1 connection */
    'id'?: number;
  };
// ...
};